Add Bayesian blocks discretization
Add new Bayesian blocks linear discretizer. Bayesian blocks is
adaptive so doesn't accept a number of bins; instead it estimates
the optimal number of bins.

This is a simple implementation, with hard-coded fitness function
and prior, which could be made more generic in the future. The
algorithm can work with different data modes, but here only the
mode suitable for histograms is implemented.